pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
dependencies = [
    "httpx==0.28.1",
    "pytest-playwright==0.6.2",
    "websockets==14.2",
]
license = {file = "LICENSE"}
name = "miranda"
readme = "README.md"
requires-python = ">=3.11"
version = "20240823"
